LayoutTests/imported/w3c/web-platform-tests/event-timing/interactionid-keyboard-event-simulated-click-link-enter-expected.txt


FAIL Click event triggered by space key in checkbox should have the same      interaction id as that of the associated keyboard events. assert_equals: Click event should have the same interaction id as the keydown event. expected 1002 but got 0
